+++ mollie.php
@@ -12,6 +12,7 @@
 
 use Mollie\Config\Config;
 use Mollie\Repository\PaymentMethodRepositoryInterface;
+use Mollie\Utility\PsVersionUtility;
 
 require_once __DIR__ . '/vendor/autoload.php';
 
@@ -48,7 +49,7 @@ class Mollie extends PaymentModule
     {
         $this->name = 'mollie';
         $this->tab = 'payments_gateways';
-        $this->version = '5.1.0';
+        $this->version = '5.2.0';
         $this->author = 'Mollie B.V.';
         $this->need_instance = 1;
         $this->bootstrap = true;
@@ -365,9 +366,12 @@ class Mollie extends PaymentModule
         );
         $this->context->controller->addJS("{$this->_path}views/js/front/mollie_iframe.js");
         $this->context->controller->addJS("{$this->_path}views/js/front/mollie_single_click.js");
+        $this->context->controller->addJS("{$this->_path}views/js/front/bancontact/qr_code.js");
+        $this->context->controller->addCSS($this->getPathUri() . 'views/css/front/bancontact_qr_code.css');
 
         Media::addJsDef([
             'ajaxUrl' => $this->context->link->getModuleLink('mollie', 'ajax'),
+            'bancontactAjaxUrl' => $this->context->link->getModuleLink('mollie', 'bancontactAjax'),
         ]);
         $this->context->controller->addJS("{$this->_path}views/js/front/mollie_error_handle.js");
         $this->context->controller->addCSS("{$this->_path}views/css/mollie_iframe.css");
@@ -917,12 +921,21 @@ class Mollie extends PaymentModule
 
     public function hookDisplayProductActions($params)
     {
-        return $this->display(__FILE__, 'views/templates/front/apple_pay_direct.tpl');
+        if (PsVersionUtility::isPsVersionHigherThen(_PS_VERSION_, '1.7.6.0')) {
+            return $this->display(__FILE__, 'views/templates/front/apple_pay_direct.tpl');
+        }
     }
 
     public function hookDisplayExpressCheckout($params)
     {
         return $this->display(__FILE__, 'views/templates/front/apple_pay_direct.tpl');
+    }
+
+    public function hookDisplayProductAdditionalInfo()
+    {
+        if (!PsVersionUtility::isPsVersionHigherThen(_PS_VERSION_, '1.7.6.0')) {
+            return $this->display(__FILE__, 'views/templates/front/apple_pay_direct.tpl');
+        }
     }
 
     /**