+++ src/Install/Installer.php
@@ -173,6 +173,7 @@ class Installer implements InstallerInterface
             'displayProductActions',
             'displayExpressCheckout',
             'actionObjectOrderPaymentAddAfter',
+            'displayProductAdditionalInfo',
         ];
     }
 
@@ -426,6 +427,7 @@ class Installer implements InstallerInterface
 
         $this->configurationAdapter->updateValue(Config::MOLLIE_API, Config::MOLLIE_ORDERS_API);
         $this->configurationAdapter->updateValue(Config::MOLLIE_APPLE_PAY_DIRECT_STYLE, 0);
+        $this->configurationAdapter->updateValue(Config::MOLLIE_BANCONTACT_QR_CODE_ENABLED, 0);
     }
 
     public function setDefaultCarrierStatuses()