+++ changelog.md
@@ -6,6 +6,7 @@
 + Added QR code option for bancontact payment on live mode
 + Added in3 payment method restriction for using only order API.
 + Changed webhook logic to stop tracking sentry errors when expired transaction webhook is called.
++ phone number exception fix when number is only +
 
 ## Changes in release 5.1.0 ##
 + Added order refund statuses for methods that use Order API.