+++ changelog.md
@@ -1,6 +1,14 @@
 ![Mollie](https://www.mollie.nl/files/Mollie-Logo-Style-Small.png)
 
 # Changelog #
+## Changes in release 5.2.0 ##
++ Fixed issue where customer id was not added when using saved credit card for order API.
++ Added QR code option for bancontact payment on live mode
++ Added in3 payment method restriction for using only order API.
++ Changed webhook logic to stop tracking sentry errors when expired transaction webhook is called.
++ Phone number exception fix when number is only +
++ Fixed issue with payment fee and price display.
+
 ## Changes in release 5.1.0 ##
 + Added order refund statuses for methods that use Order API.
 + Fixed refund and cancel information text in order page.